getNavigationDetails() and getPagingNavigation moved to own CalendarNavigationService
