+++ src/main/java/com/celements/calendar/navigation/PagingNavigation.java
@@ -1,4 +1,4 @@
-package com.celements.calendar.manager;
+package com.celements.calendar.navigation;
 
 public class PagingNavigation {
