+++ src/test/java/com/celements/calendar/manager/EventsManagerTest.java
@@ -16,7 +16,6 @@ import org.xwiki.model.reference.SpaceReference;
 import org.xwiki.model.reference.WikiReference;
 
 import com.celements.calendar.Calendar;
-import com.celements.calendar.Event;
 import com.celements.calendar.ICalendar;
 import com.celements.calendar.IEvent;
 import com.celements.calendar.engine.ICalendarEngineRole;
@@ -25,7 +24,6 @@ import com.celements.common.test.AbstractBridgedComponentTestCase;
 import com.xpn.xwiki.XWiki;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
-import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.web.Utils;
 
 public class EventsManagerTest extends AbstractBridgedComponentTestCase {
@@ -138,83 +136,6 @@ public class EventsManagerTest extends AbstractBridgedComponentTestCase {
     verifyAll();
 
     assertEquals(12345L, countEvent);
-  }
-
-  @Test
-  public void testGetNavigationDetails_noStartDate() throws XWikiException {
-    DocumentReference cal1Ref = new DocumentReference(context.getDatabase(), "mySpace",
-        "cal1");
-    Calendar cal = new Calendar(cal1Ref, false);
-    Event event = createMock(Event.class);
-    DocumentReference eventRef = new DocumentReference(context.getDatabase(),
-        "myEventSpace", "Event1");
-    XWikiDocument cal1Doc = new XWikiDocument(cal1Ref);
-
-    expect(event.getEventDate()).andReturn(null).atLeastOnce();
-    expect(calServiceMock.getMidnightDate((Date) eq(null))).andReturn(null);
-    expect(event.getDocumentReference()).andReturn(eventRef).anyTimes();
-    expect(xwiki.getDocument(eq(cal1Ref), same(context))).andReturn(cal1Doc
-        ).anyTimes();
-
-    replayAll(event);
-    assertNull(eventsMgr.getNavigationDetails(event, cal));
-    verifyAll(event);
-  }
-
-  @Test
-  public void testGetNavigationDetails() throws XWikiException {
-    List<String> spaces = Arrays.asList("asdf");
-    EntityReference calSpace = new SpaceReference(spaces.get(0), new WikiReference(
-        context.getDatabase()));
-    Date eventDate = new Date();
-    Date eventMidnightDate = getCalService().getMidnightDate(eventDate);
-    IEvent event = createMock(IEvent.class);
-    DocumentReference eventDocRef = new DocumentReference(context.getDatabase(),
-        "mySpace", "myEvent");
-    eventDocRef.setParent(calSpace);
-    IEvent event2 = createMock(IEvent.class);
-    DocumentReference eventDocRef2 = new DocumentReference(context.getDatabase(),
-        "mySpace", "myEvent2");
-    eventDocRef2.setParent(calSpace);
-    IEvent event3 = createMock(IEvent.class);
-    DocumentReference eventDocRef3 = new DocumentReference(context.getDatabase(),
-        "mySpace", "myEvent3");
-    eventDocRef3.setParent(calSpace);
-    List<IEvent> eventList = Arrays.asList(event2, event3, event);
-    DocumentReference calDocRef = new DocumentReference(context.getDatabase(), "mySpace",
-        "myCalDoc");
-    Date startDate = new Date(1000);
-    String lang = "de";
-    ICalendar calMock = createMock(Calendar.class);
-
-    expect(calMock.getDocumentReference()).andReturn(calDocRef).anyTimes();
-    expect(calMock.isArchive()).andReturn(false).anyTimes();
-    expect(calMock.getStartDate()).andReturn(startDate).anyTimes();
-    expect(calMock.getEngine()).andReturn(engineMock).anyTimes();
-    expect(event.getEventDate()).andReturn(eventDate).once();
-    expect(calServiceMock.getMidnightDate(eq(eventDate))).andReturn(eventMidnightDate
-        ).once();
-
-    expect(xwiki.getSpacePreference(eq("default_language"), same(context))
-        ).andReturn(lang);
-    expect(calServiceMock.getAllowedSpaces(eq(calDocRef))).andReturn(spaces).once();
-    expect(engineMock.getEvents(eq(eventMidnightDate), eq(false), eq(lang), eq(spaces),
-        eq(0), eq(10))).andReturn(eventList).once();
-
-    expect(calServiceMock.getEventSpaceForCalendar(eq(calDocRef))).andReturn(spaces.get(0)
-        ).times(3);
-    expect(event.getDocumentReference()).andReturn(eventDocRef).once();
-    expect(event2.getDocumentReference()).andReturn(eventDocRef2).once();
-    expect(event3.getDocumentReference()).andReturn(eventDocRef3).once();
-
-    replayAll(calMock, event, event2, event3);
-    NavigationDetails navDetails = eventsMgr.getNavigationDetails(event, calMock);
-    verifyAll(calMock, event, event2, event3);
-
-    assertNotNull(navDetails);
-    assertEquals(eventMidnightDate, navDetails.getStartDate());
-    assertEquals(2, navDetails.getOffset());
-
   }
 
   @Test
